+++ src/ffi.rs
@@ -2,28 +2,34 @@ use crate::Errno;
 use crate::Sysno;
 
 extern "C" {
-    fn __syscall0(nr: i64) -> i64;
-    fn __syscall1(nr: i64, arg1: u64) -> i64;
-    fn __syscall2(nr: i64, arg1: u64, arg2: u64) -> i64;
-    fn __syscall3(nr: i64, arg1: u64, arg2: u64, arg3: u64) -> i64;
-    fn __syscall4(nr: i64, arg1: u64, arg2: u64, arg3: u64, arg4: u64) -> i64;
+    fn __syscall0(nr: usize) -> usize;
+    fn __syscall1(nr: usize, arg1: usize) -> usize;
+    fn __syscall2(nr: usize, arg1: usize, arg2: usize) -> usize;
+    fn __syscall3(nr: usize, arg1: usize, arg2: usize, arg3: usize) -> usize;
+    fn __syscall4(
+        nr: usize,
+        arg1: usize,
+        arg2: usize,
+        arg3: usize,
+        arg4: usize,
+    ) -> usize;
     fn __syscall5(
-        nr: i64,
-        arg1: u64,
-        arg2: u64,
-        arg3: u64,
-        arg4: u64,
-        arg5: u64,
-    ) -> i64;
+        nr: usize,
+        arg1: usize,
+        arg2: usize,
+        arg3: usize,
+        arg4: usize,
+        arg5: usize,
+    ) -> usize;
     fn __syscall6(
-        nr: i64,
-        arg1: u64,
-        arg2: u64,
-        arg3: u64,
-        arg4: u64,
-        arg5: u64,
-        arg6: u64,
-    ) -> i64;
+        nr: usize,
+        arg1: usize,
+        arg2: usize,
+        arg3: usize,
+        arg4: usize,
+        arg5: usize,
+        arg6: usize,
+    ) -> usize;
 }
 
 /// Issues a system call with 0 arguments.
@@ -33,8 +39,8 @@ extern "C" {
 /// Running a system call is inherently unsafe. It is the caller's
 /// responsibility to ensure safety.
 #[inline(always)]
-pub unsafe fn syscall0(nr: Sysno) -> Result<i64, Errno> {
-    Errno::from_ret(__syscall0(nr as i64))
+pub unsafe fn syscall0(nr: Sysno) -> Result<usize, Errno> {
+    Errno::from_ret(__syscall0(nr as usize))
 }
 
 /// Issues a system call with 1 arguments.
@@ -44,8 +50,8 @@ pub unsafe fn syscall0(nr: Sysno) -> Result<i64, Errno> {
 /// Running a system call is inherently unsafe. It is the caller's
 /// responsibility to ensure safety.
 #[inline(always)]
-pub unsafe fn syscall1(nr: Sysno, a1: u64) -> Result<i64, Errno> {
-    Errno::from_ret(__syscall1(nr as i64, a1))
+pub unsafe fn syscall1(nr: Sysno, a1: usize) -> Result<usize, Errno> {
+    Errno::from_ret(__syscall1(nr as usize, a1))
 }
 
 /// Issues a system call with 2 arguments.
@@ -55,8 +61,12 @@ pub unsafe fn syscall1(nr: Sysno, a1: u64) -> Result<i64, Errno> {
 /// Running a system call is inherently unsafe. It is the caller's
 /// responsibility to ensure safety.
 #[inline(always)]
-pub unsafe fn syscall2(nr: Sysno, a1: u64, a2: u64) -> Result<i64, Errno> {
-    Errno::from_ret(__syscall2(nr as i64, a1, a2))
+pub unsafe fn syscall2(
+    nr: Sysno,
+    a1: usize,
+    a2: usize,
+) -> Result<usize, Errno> {
+    Errno::from_ret(__syscall2(nr as usize, a1, a2))
 }
 
 /// Issues a system call with 3 arguments.
@@ -68,11 +78,11 @@ pub unsafe fn syscall2(nr: Sysno, a1: u64, a2: u64) -> Result<i64, Errno> {
 #[inline(always)]
 pub unsafe fn syscall3(
     nr: Sysno,
-    a1: u64,
-    a2: u64,
-    a3: u64,
-) -> Result<i64, Errno> {
-    Errno::from_ret(__syscall3(nr as i64, a1, a2, a3))
+    a1: usize,
+    a2: usize,
+    a3: usize,
+) -> Result<usize, Errno> {
+    Errno::from_ret(__syscall3(nr as usize, a1, a2, a3))
 }
 
 /// Issues a system call with 4 arguments.
@@ -84,12 +94,12 @@ pub unsafe fn syscall3(
 #[inline(always)]
 pub unsafe fn syscall4(
     nr: Sysno,
-    a1: u64,
-    a2: u64,
-    a3: u64,
-    a4: u64,
-) -> Result<i64, Errno> {
-    Errno::from_ret(__syscall4(nr as i64, a1, a2, a3, a4))
+    a1: usize,
+    a2: usize,
+    a3: usize,
+    a4: usize,
+) -> Result<usize, Errno> {
+    Errno::from_ret(__syscall4(nr as usize, a1, a2, a3, a4))
 }
 
 /// Issues a system call with 5 arguments.
@@ -101,13 +111,13 @@ pub unsafe fn syscall4(
 #[inline(always)]
 pub unsafe fn syscall5(
     nr: Sysno,
-    a1: u64,
-    a2: u64,
-    a3: u64,
-    a4: u64,
-    a5: u64,
-) -> Result<i64, Errno> {
-    Errno::from_ret(__syscall5(nr as i64, a1, a2, a3, a4, a5))
+    a1: usize,
+    a2: usize,
+    a3: usize,
+    a4: usize,
+    a5: usize,
+) -> Result<usize, Errno> {
+    Errno::from_ret(__syscall5(nr as usize, a1, a2, a3, a4, a5))
 }
 
 /// Issues a system call with 6 arguments.
@@ -119,12 +129,12 @@ pub unsafe fn syscall5(
 #[inline(always)]
 pub unsafe fn syscall6(
     nr: Sysno,
-    a1: u64,
-    a2: u64,
-    a3: u64,
-    a4: u64,
-    a5: u64,
-    a6: u64,
-) -> Result<i64, Errno> {
-    Errno::from_ret(__syscall6(nr as i64, a1, a2, a3, a4, a5, a6))
+    a1: usize,
+    a2: usize,
+    a3: usize,
+    a4: usize,
+    a5: usize,
+    a6: usize,
+) -> Result<usize, Errno> {
+    Errno::from_ret(__syscall6(nr as usize, a1, a2, a3, a4, a5, a6))
 }
